# Improved performance of call_and_shelve