@aabadie 2018-08-24
Nice one! It would be great to also add a test.