@MaximeWeyl 2018-08-24
Sure I could write one. I have been thinking about what should be tested, but did not find any good idea.

Indeed, this PR does not change, add, nor removes any functionality. Just a tiny bit of the implementation of the current functionalities. 

I could test that the performance remains below a given threshold, but I'm not so sure it is a good idea (mainly because it depends on the worker of the test). 

I could also mock the *store_backend* and assert that it has not been called, but : it depends on the implementation, and I have not been able to find any use of mocking in the entire repository. It is the best idea I was able to come up with nonetheless. 

What do you think about it ?