@aabadie 2018-08-24
> I could also mock the store_backend and assert that it has not been called, but : it depends on the implementation, and I have not been able to find any use of mocking in the entire repository. It is the best idea I was able to come up with nonetheless.

I had the same idea and that's what I would also do. But maybe other maintainers have better ideas. @lesteve @ogrisel @tomMoral, what do you think ?