@ogrisel 2018-08-27
LGTM, merging. Thank you very much for this contribution @MaximeWeyl !